+++ src/main/java/com/github/cao/awa/apricot/network/packet/receive/message/MessagePacket.java
@@ -16,7 +16,7 @@ public abstract class MessagePacket extends ReadonlyPacket {
 
     public abstract long getTimestamp();
 
-    public abstract int getMessageId();
+    public abstract long getMessageId();
 
     public abstract long getBotId();
